Sphere
A computer program receives two inputs. First input is circle’s radius. Second input is operation selection; circumference or area of circle. If selection input is ‘c’ or ‘C’, circumference of the circle is computed and displayed. Else if selection input is ‘a’ or ‘A’, area of the circle is computed and displayed. Else if selection input is ‘v’ or ‘V’, volume of sphere is computed and displayed. Else a message that says “invalid selection.” Is displayed.